+++ engine/physics/__init__.py
@@ -1,3 +1,3 @@
-from .two_dimensional_physics import Physics2d
+from .physics_2d import Physics2d
 
 __all__ = ['Physics2d']
