+++ engine/physics/tests/test_physics_2d.py
@@ -1,4 +1,4 @@
-from ..two_dimensional_physics import Physics2d
+from ..physics_2d import Physics2d
 import unittest
